backend/requirements.txt

fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
sqlalchemy>=1.4.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.19.0
python-dotenv>=0.19.0
requests>=2.26.0
pyjwt>=2.1.0
pytest>=6.2.5
httpx>=0.19.0
supabase>=0.7.1
fastapi-cors>=0.0.6
python-dateutil>=2.8.2
